Add CartController and update Item toString method
A new controller named CartController is added to manage cart operations. The toString method in the Item class is updated to use Gson for serialization instead of the previous implementation. This will offer more flexibility and control over the JSON output. Also, the API routing in InventoryController is updated, and necessary dependencies are added in pom.xml.