Add build and docker scripts and update Dockerfile
New scripts for building the backend application and running it in a Docker container have been added. The Dockerfile was updated to include the necessary dependencies for the build, and the application name in multiple scripts was corrected to 'pos_backend'. Additionally, added '.secrets' to .gitignore to secure sensitive information.